refactor: 브릿지 코드 정리
