--- src/bridge/ipc/exception/bridge-not-initialized.ts
@@ -1,5 +1,0 @@
-export class BridgeNotInitialized extends Error {
-  constructor() {
-    super('Bridge not initialized');
-  }
-}
