+++ src/bridge/ipc/rawIpc/RawIpcFactory.ts
@@ -1,31 +1,17 @@
 import { AbstractRawIpc } from '@/bridge/ipc/rawIpc/AbstractRawIpc';
-import { FlutterWebviewIpc } from '@/bridge/ipc/rawIpc/flutter/FlutterWebviewIpc';
+import { FlutterWebviewIpc } from '@/bridge/ipc/rawIpc/FlutterWebviewIpc';
 
 export class RawIpcFactory {
   static async create(): Promise<AbstractRawIpc | null> {
     console.log('create called');
     try {
-      const flutterWebviewIpc = await retry(3, () => {
+      return await retry(3, () => {
         return new FlutterWebviewIpc();
       });
-      
-      return flutterWebviewIpc;
       
     } catch (error) {
       console.error('Failed to create FlutterWebviewIpc', error);
       return null;
-    }
-  }
-  
-  static async isFlutterWebviewAvailable(): Promise<boolean> {
-    try {
-      await retry(3, () => {
-        return new FlutterWebviewIpc();
-      });
-      
-      return true;
-    } catch (error) {
-      return false;
     }
   }
 }
@@ -37,7 +23,6 @@ const retry = <T>(maxTries: number, fn: () => T): Promise<T> => {
     const tryFn = () => {
       tries++;
       
-      console.log('tryFn called');
       try {
         resolve(fn());
       } catch (error) {
